# docs/battleships-api-spec.yaml
openapi: 3.0.3
info:
  title: Battleships API
  version: 0.0.1
  description: Web-based system that allows multiple players to play the Battleship game.
servers:
  - url: http://localhost:8888
    description: Local server
...
paths:
...
  /games:
    get:
      tags:
        - Games
      summary: Gets all games
      operationId: getGames
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        200:
          description: Games were found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetGamesOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Games
      summary: Create a game
      operationId: createGame
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Created game object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameInputModel'
      responses:
        201:
          description: Game was created
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/CreateGameOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/matchmake:
    post:
      tags:
        - Games
      summary: Matchmake a game
      operationId: matchmake
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Configuration of the game to be matched
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameConfigModel'
      responses:
        200:
          description: Game was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/MatchmakeOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Gets a game by id
      operationId: getGame
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Game was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetGameOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/state:
    get:
      tags:
        - Games
      summary: Gets a game state by id
      operationId: getGameState
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Game state was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetGameStateOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/join:
    post:
      tags:
        - Games
      summary: Join a game
      operationId: joinGame
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: User joined the game
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/JoinGameOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/players/self/fleet:
    get:
      tags:
        - Players
      summary: Gets the player's fleet
      operationId: getFleet
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Fleet was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetFleetOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Players
      summary: Deploy a fleet
      operationId: deployFleet
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      requestBody:
        description: Fleet object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployFleetInputModel'
      responses:
        200:
          description: Fleet was deployed
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/DeployFleetOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /games/{gameId}/players/opponent/fleet:
    get:
      tags:
        - Players
      summary: Gets the opponent's fleet
      operationId: getOpponentFleet
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Fleet was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetOpponentFleetOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/players/self/shots:
    get:
      tags:
        - Players
      summary: Gets the player's shots
      operationId: getShots
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Shots were found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetShotsOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Players
      summary: Fire shots
      operationId: fireShots
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      requestBody:
        description: Shots object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FireShotsInputModel'
      responses:
        201:
          description: Fleet was deployed
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/FireShotsOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}/players/opponent/shots:
    get:
      tags:
        - Players
      summary: Gets the opponent's shots
      operationId: getOpponentShots
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/gameIdParam'
      responses:
        200:
          description: Shots were found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetOpponentShotsOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    PlayerModel:
      type: object
      properties:
        username:
          type: string
          example: guest1
        points:
          type: integer
          example: 100
    GameStateModel:
      type: object
      properties:
        phase:
          type: string
          example: WAITING_FOR_PLAYERS
        phaseEndTime:
          type: integer
          example: 1666201815147
        round:
          type: integer
          example: 1
        turn:
          type: string
          example: guest1
        winner:
          type: string
          example: guest1
    GameConfigModel:
      type: object
      required:
        - gridSize
        - maxTimeForLayoutPhase
        - shotsPerRound
        - maxTimePerRound
        - shipTypes
      properties:
        gridSize:
          type: integer
          example: 10
        maxTimeForLayoutPhase:
          type: integer
          example: 60
        shotsPerRound:
          type: integer
          example: 5
        maxTimePerRound:
          type: integer
          example: 40
        shipTypes:
          type: array
          items:
            $ref: '#/components/schemas/ShipTypeModel'
    MatchmakeOutputModel:
      type: object
      properties:
        wasCreated:
          type: boolean
          example: true
    JoinGameOutputModel:
      type: object
      properties:
        gameId:
          type: integer
          example: 1
    GetGameStateOutputModel:
      $ref: '#/components/schemas/GameStateModel'
    GetGamesOutputModel:
      type: object
      properties:
        totalCount:
          type: integer
          example: 10
    GetGameOutputModel:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: game1
        creator:
          type: string
          example: guest1
        config:
          $ref: '#/components/schemas/GameConfigModel'
        state:
          $ref: '#/components/schemas/GameStateModel'
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerModel'
    CreateGameInputModel:
      type: object
      required:
        - name
        - config
      properties:
        name:
          type: string
          example: game1
        config:
          $ref: '#/components/schemas/GameConfigModel'
    DeployedShipModel:
      type: object
      properties:
        type:
          type: string
          example: DESTROYER
        coordinate:
          $ref: '#/components/schemas/CoordinateModel'
        orientation:
          type: string
          example: HORIZONTAL
        lives:
          type: integer
          example: 2
    GetShotsOutputModel:
      $ref: '#/components/schemas/FiredShotsModel'
    GetOpponentShotsOutputModel:
      $ref: '#/components/schemas/FiredShotsModel'
    GetOpponentFleetOutputModel:
      $ref: '#/components/schemas/GetFleetOutputModel'
    GetFleetOutputModel:
      type: object
      properties:
        ships:
          type: array
          items:
            $ref: '#/components/schemas/DeployedShipModel'
...
    FireShotsOutputModel:
      $ref: '#/components/schemas/FiredShotsModel'
    FireShotsInputModel:
      type: object
      required:
        - shots
      properties:
        shots:
          type: array
          items:
            $ref: '#/components/schemas/UnfiredShotModel'
    UnfiredShotModel:
      type: object
      required:
        - coordinate
      properties:
        coordinate:
          $ref: '#/components/schemas/CoordinateModel'
    ShotResultModel:
      type: object
      properties:
        result:
          type: string
          example: HIT
    FiredShotsModel:
      type: object
      properties:
        shots:
          type: array
          items:
            $ref: '#/components/schemas/FiredShotModel'
    FiredShotModel:
      type: object
      properties:
        coordinate:
          $ref: '#/components/schemas/CoordinateModel'
        round:
          type: integer
          example: 1
        result:
          $ref: '#/components/schemas/ShotResultModel'
    DeployFleetOutputModel:
      type: object
      properties:
        successfullyDeployed:
          type: boolean
          example: true
    DeployFleetInputModel:
      type: object
      required:
        - ships
      properties:
        ships:
          type: array
          items:
            $ref: '#/components/schemas/UndeployedShipModel'
    UndeployedShipModel:
      type: object
      required:
        - type
        - coordinate
        - orientation
      properties:
        type:
          type: string
          example: DESTROYER
        coordinate:
          $ref: '#/components/schemas/CoordinateModel'
        orientation:
          type: string
          example: HORIZONTAL
    ShipTypeModel:
      type: object
      required:
        - shipName
        - quantity
        - size
        - points
      properties:
        shipName:
          type: string
          example: DESTROYER
        quantity:
          type: integer
          example: 1
        size:
          type: integer
          example: 2
        points:
          type: integer
          example: 2
    CoordinateModel:
      type: object
      required:
        - col
        - row
      properties:
        col:
          type: string
          example: A
        row:
          type: integer
          example: 1

    # Siren Base Models
    Entity:
      type: object
      properties:
        class:
          type: array
          items:
            type: string
        properties:
          type: object
          additionalProperties: false
        entities:
          type: array
          items:
            $ref: '#/components/schemas/SubEntity'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    SubEntity:
      type: object
      required:
        - rel
      properties:
        class:
          type: array
          items:
            type: string
        rel:
          type: array
          items:
            type: string
        href:
          type: string
        type:
          type: string
        title:
          type: string
        properties:
          type: object
          additionalProperties: false
        entities:
          type: array
          items:
            $ref: '#/components/schemas/SubEntity'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Action:
      type: object
      required:
        - name
        - href
      properties:
        name:
          type: string
        title:
          type: string
        method:
          type: string
        href:
          type: string
        type:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
    Field:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        title:
          type: string
    Link:
      type: object
      required:
        - rel
        - href
      properties:
        rel:
          type: array
          items:
            type: string
        href:
          type: string
        type:
          type: string
        title:
          type: string
    ErrorModel:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        path:
          type: string
  parameters:
    offsetParam:
      name: offset
      in: query
      description: The offset to start from
      required: true
      schema:
        type: integer
        default: 0
    limitParam:
      name: limit
      in: query
      description: The number of items to return
      required: true
      schema:
        type: integer
        default: 10
...
    gameIdParam:
      name: gameId
      in: path
      description: The id of the game
      required: true
      schema:
        type: integer
        example: 1
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT